requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiostream==0.5.2
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.0.339
langsmith==0.0.92
litellm==1.2.0
llama-index==0.9.4
llmsherpa==0.1.4
lyzr==0.1.30
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.3.4
packaging==23.2
pandas==2.0.2
pdfminer.six
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
streamlit==1.32.2
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
validators==0.22.0
watchdog==4.0.0
weaviate-client==3.25.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
